Refacto team list + core headers
