+++ src/server/CMakeLists.txt
@@ -12,6 +12,7 @@ target_include_directories(zappy_server PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/cor
 target_include_directories(zappy_server PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/implementation")
 target_include_directories(zappy_server PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/utils")
 target_include_directories(zappy_server PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/include")
+target_include_directories(zappy_server PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/include/core")
 target_include_directories(zappy_server PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/include/implementation")
 
 set_target_properties(zappy_server
