+++ src/server/core/CMakeLists.txt
@@ -3,7 +3,6 @@ set(CMAKE_CXX_FLAGS "-Wall -Wextra -Wpedantic")
 target_sources(zappy_server PRIVATE
         "${CMAKE_CURRENT_SOURCE_DIR}/handle_clients.c"
         "${CMAKE_CURRENT_SOURCE_DIR}/launch_server.c"
-        "${CMAKE_CURRENT_SOURCE_DIR}/server_data.c"
         "${CMAKE_CURRENT_SOURCE_DIR}/detect_win.c"
         "${CMAKE_CURRENT_SOURCE_DIR}/read_client.c"
         "${CMAKE_CURRENT_SOURCE_DIR}/clients_lifecycle.c"
@@ -15,8 +14,10 @@ target_sources(zappy_server PRIVATE
         )
 
 add_subdirectory(options)
+add_subdirectory(server_data)
 
 target_include_directories(zappy_server PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/options")
+target_include_directories(zappy_server PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/server_data")
 
 set_target_properties(zappy_server
     PROPERTIES
