+++ src/server/core/options/option_setters.c
@@ -5,9 +5,11 @@
 ** option_setters
 */
 
-#include <stdio.h>
 #include "core.h"
 #include "utils.h"
+#include "server_options.h"
+#include "server_data.h"
+#include "default_values.h"
 
 int set_port(data_t *data, char *value)
 {
@@ -17,7 +19,22 @@ int set_port(data_t *data, char *value)
 
 int set_team_names(data_t *data, char *value)
 {
-    data->team_names = str_to_word_array(value, " ");
+    char **team_names = NULL;
+    int nb_teams = 0;
+
+    free_teams(data);
+    data->teams = NULL;
+    team_names = str_to_word_array(value, " ");
+    nb_teams = word_array_len(team_names);
+    data->teams = malloc(sizeof(team_t *) * (nb_teams + 1));
+    for (int i = 0; i < nb_teams; i++) {
+        data->teams[i] = malloc(sizeof(team_t));
+        data->teams[i]->name = strdup(team_names[i]);
+        data->teams[i]->nb_cli = DEFAULT_CLI_PER_TEAM;
+        data->teams[i]->fork_mode = 0;
+    }
+    data->teams[nb_teams] = NULL;
+    free_word_array(team_names);
     return 0;
 }
 
@@ -29,7 +46,9 @@ int set_cli_per_team(data_t *data, char *value)
         print_help(data, NULL);
         return 84;
     }
-    data->cli_per_team = atoi(value);
+    for (int i = 0; data->teams[i]; i++) {
+        data->teams[i]->nb_cli = atoi(value);
+    }
     return 0;
 }
