+++ src/server/implementation/gui/cmd/tna.c
@@ -15,9 +15,9 @@ int do_tna(data_t *data, char **args)
 
     if (args)
         return 1;
-    for (int i = 0; data->team_names[i]; i++) {
+    for (int i = 0; data->teams[i]; i++) {
         result = strdup("tna ");
-        result = concat_str(result, data->team_names[i]);
+        result = concat_str(result, data->teams[i]->name);
         result = concat_str(result, "\n");
         send_to_client(data->clients, data->curr_cli_index, result);
         free(result);
