+++ src/server/implementation/input_parser.c
@@ -16,11 +16,11 @@ static int parse_team_name(data_t *data)
         data->clients[data->curr_cli_index]->is_registered = 1;
         return append_gui_client(data);
     }
-    for (int i = 0; data->team_names[i]; i++) {
-        if (!strcmp(data->team_names[i],
+    for (int i = 0; data->teams[i]; i++) {
+        if (!strcmp(data->teams[i]->name,
             data->clients[data->curr_cli_index]->input)) {
             data->clients[data->curr_cli_index]->is_registered = 1;
-            return append_player_client(data, data->team_names[i]);
+            return append_player_client(data, data->teams[i]->name);
         }
     }
     send_to_client(data->clients, data->curr_cli_index, "ko\n");
