+++ src/server/include/default_values.h
@@ -1,0 +1,24 @@
+/*
+** EPITECH PROJECT, 2023
+** zappy_server
+** File description:
+** Header for default values
+*/
+
+#ifndef ZAPPY_DEFAULT_VALUES_H
+    #define ZAPPY_DEFAULT_VALUES_H
+
+    #define DEFAULT_PORT 4242
+    #define DEFAULT_IP_ADDRESS "127.0.0.1"
+
+    #define DEFAULT_MAP_WIDTH 10
+    #define DEFAULT_MAP_HEIGHT 10
+
+    #define DEFAULT_FREQ 100
+
+    #define LEVEL_START 1
+
+    #define DEFAULT_CLI_PER_TEAM 2
+    #define DEFAULT_TEAM_NAME "Team"
+
+#endif //ZAPPY_DEFAULT_VALUES_H
